+++ src/plugins/optionsHandler/index.ts
@@ -1,1 +1,2 @@
+export * from './noOptionsHandler/noOptionsHandler.component';
 export * from './simpleOptionsHandler/simpleOptionsHandler.component';